requirements.txt
asgiref==3.7.2
attrs==23.2.0
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==4.2.8
django-bootstrap5==23.4
django-filter==23.5
django-summernote==0.8.20.0
django-taggit==3.0.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.2.0
psycopg==3.1.17
psycopg-binary==3.1.17
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
webencodings==0.5.1
